moved the new endpoints to a new file in the controller
